# OPENAPI.yaml
...
openapi: 3.1.0
info:
  title: netflat API
  description: |
    REST API cho ứng dụng xem video tự sản xuất theo trải nghiệm Netflix.
    
    ## Authentication
    API sử dụng JWT Bearer token. Access token có thời hạn 15 phút, refresh token 7 ngày.
    
    ## Error Format
    Tất cả lỗi trả về theo format:
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message",
        "details": null,
        "requestId": "uuid"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: netflat Team
    email: dev@netflat.local
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://api.netflat.example.com/api
    description: Staging (placeholder)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
          properties:
            code:
              type: string
              example: VALIDATION_FAILED
            message:
              type: string
              example: Validation failed
            details:
              type: object
              nullable: true
              additionalProperties: true
            requestId:
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000

    # ─────────────────────────────────────────────────────────────
    # Auth
    # ─────────────────────────────────────────────────────────────
...
    MovieListItem:
      type: object
      description: Simplified movie for list/rail display
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        posterUrl:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        releaseYear:
          type: integer
          nullable: true
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
...
    Genre:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        name:
          type: string
          example: Action
        slug:
          type: string
          example: action
...
    Favorite:
      type: object
      properties:
        id:
          type: string
          format: uuid
        movieId:
          type: string
          format: uuid
        movie:
          $ref: '#/components/schemas/MovieListItem'
        createdAt:
          type: string
          format: date-time
    FavoriteListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Favorite'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # ─────────────────────────────────────────────────────────────
    # Upload
    # ─────────────────────────────────────────────────────────────
...
    WatchHistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        movieId:
          type: string
          format: uuid
        progressSeconds:
          type: integer
        durationSeconds:
          type: integer
        completed:
          type: boolean
        lastWatchedAt:
          type: string
          format: date-time
        movie:
          type: object
          properties:
            id: { type: string, format: uuid }
            title: { type: string }
            posterUrl: { type: string, nullable: true }
...
paths:
...
  /favorites:
    get:
      tags: [Favorites]
      summary: List favorites
      description: Get user's favorite movies (My List)
      operationId: listFavorites
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /favorites/{movieId}:
    post:
      tags: [Favorites]
      summary: Add to favorites
      description: Add a movie to user's My List
      operationId: addFavorite
      security:
        - bearerAuth: []
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Added to favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Favorite'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Already in favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: FAVORITE_ALREADY_EXISTS
                  message: "Movie already in favorites"
                  details: null
                  requestId: "550e8400-e29b-41d4-a716-446655440000"

    delete:
      tags: [Favorites]
      summary: Remove from favorites
      description: Remove a movie from user's My List
      operationId: removeFavorite
      security:
        - bearerAuth: []
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Removed from favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Removed from favorites"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not in favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─────────────────────────────────────────────────────────────
  # Upload
  # ─────────────────────────────────────────────────────────────
...
  /history:
    get:
      tags: [History]
      summary: List watch history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated watch history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WatchHistoryItem'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /history/continue-watching:
    get:
      tags: [History]
      summary: Get continue watching items (incomplete, recent)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: List of in-progress movies
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WatchHistoryItem'
  /history/{movieId}:
    post:
      tags: [History]
      summary: Upsert watch progress
      security:
        - bearerAuth: []
      parameters:
        - name: movieId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [progressSeconds]
              properties:
                progressSeconds:
                  type: integer
                  minimum: 0
                durationSeconds:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Progress saved
    delete:
      tags: [History]
      summary: Remove history entry
      security:
        - bearerAuth: []
      parameters:
        - name: movieId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: History entry removed

  # ─────────────────────────────────────────────────────────
  # Ratings
  # ─────────────────────────────────────────────────────────
...